seperate manipulation exception
