+++ code/arlab_manipulation_cpp/src/hand_motion.cpp
@@ -1,5 +1,5 @@
 #include "arlab_manipulation_cpp/hand_motion.hpp"
-#include "arlab_manipulation_cpp/orchestrator_listener.hpp"
+#include "arlab_manipulation_cpp/manipulator_exception.hpp"
 
 #include <stdexcept>
